task2

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -17,7 +17,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 
     va_start(args, n);
 
-    for (i = 0; i < n - 1; i++)
+    for (i = 0; i < (int)n - 1; i++)
     {
         printInt(va_arg(args, int));
         fputs(separator, stdout);
